pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai4sec-application-layer"
version = "0.1.0"
description = "AI4SEC application-layer delivery assets for Dify, OpenCTI MCP, and Notification MCP."
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.10.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]
